Recognizes 3-4 patterns

+++ AANN/weightmatrix.cpp
@@ -5,6 +5,6 @@ WeightMatrix::WeightMatrix()
 }
 
 WeightMatrix::WeightMatrix(const unsigned long iSize)
-    : weights(iSize, vector<int32_t>(iSize))
+    : weights(iSize, vector<double>(iSize))
 {
 }
